pyproject.toml
[tool.poetry]
name = "article-whatsapp-agent"
version = "0.1.0"
description = ""
authors = ["lukaszkowejsza <lukasz@deyan7.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langfuse = "^2.33.1"
openai = "^1.30.5"
langchain = "^0.2.1"
llama-index = "^0.10.42"
anthropic = "^0.28.0"
colorama = "^0.4.6"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
jupyter = "^1.0.0"
sqlmodel = "^0.0.18"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.3"
psycopg2 = "^2.9.9"
pydub = "^0.25.1"
firebase-admin = "^6.5.0"
gcloud = "^0.18.3"
pycryptodomex = "^3.20.0"
pycryptodome = "^3.20.0"
uvicorn = "0.30.1"
retry = "0.9.2"
sse-starlette = "1.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
